+++ www/install/clearcache.php
@@ -1,0 +1,8 @@
+<?php
+
+use go\core\App;
+
+require('../vendor/autoload.php');
+App::get();
+
+go()->rebuildCache();