+++ CHANGELOG.md
@@ -22,6 +22,8 @@ XX-09-20216.6.10
 
 
 
+- Core: Clear cache via webserver if running upgrade on the command line.
+
 09-12-2021 6.5.92
 - ActiveSync: Reverted Michael z-push calendar stuff because it's unfinished and causing problems